+++ src/core/queue/interface.ts
@@ -9,10 +9,14 @@
 export interface InnerQueue<T = unknown> {
 	readonly length: number;
 	readonly head?: CanUndef<T>;
+	readonly first?: CanUndef<T>;
 	push(task: T): number;
 	unshift(task: T): number;
 	pop(): CanUndef<T>;
 	shift(): CanUndef<T>;
+	clear?(): void;
+	clone?(): InnerQueue<T>;
+	[Symbol.iterator](): IterableIterator<T>;
 }
 
 export interface CreateInnerQueue<T extends InnerQueue<any> = InnerQueue> {
@@ -81,7 +85,7 @@ export default abstract class Queue<T> {
 	/**
 	 * Returns iterator
 	 */
-	[Symbol.iterator](): IterableIterator<unknown> {
+	[Symbol.iterator](): IterableIterator<T> {
 		const
 			clonedQueue = this.clone();
 
@@ -90,12 +94,16 @@ export default abstract class Queue<T> {
 				return this;
 			},
 
-			next(): IteratorResult<unknown> {
+			next(): IteratorResult<T> {
 				const
 					done = clonedQueue.length <= 0,
 					value = clonedQueue.pop();
 
-				return {value, done};
+				if (done || value == null) {
+					return {done: true, value: undefined};
+				}
+
+				return {done, value};
 			}
 		};
 	}
@@ -103,7 +111,7 @@ export default abstract class Queue<T> {
 	/**
 	 * Returns async iterator
 	 */
-	[Symbol.asyncIterator](): AsyncIterableIterator<unknown> {
+	[Symbol.asyncIterator](): AsyncIterableIterator<T> {
 		const
 			clonedQueue = this.clone();
 
@@ -112,12 +120,18 @@ export default abstract class Queue<T> {
 				return this;
 			},
 
-			next(): Promise<IteratorResult<unknown>> {
-				const
-					done = clonedQueue.length <= 0,
-					value = clonedQueue.pop();
+			next(): Promise<IteratorResult<T>> {
+				return new Promise((resolve) => {
+					const
+						done = clonedQueue.length <= 0,
+						value = clonedQueue.pop();
 
-				return Promise.resolve({value, done});
+					if (done || value == null) {
+						return resolve({done: true, value: undefined});
+					}
+
+					return resolve({done, value});
+				});
 			}
 		};
 	}